revamp folder structure and build+dist process
