+++ README.md
@@ -1,5 +1,9 @@
-# ipfs-electron
-Electron Shell based ipfs app
+IPFS Native Application
+=======================
+
+[![](https://img.shields.io/badge/made%20by-Protocol%20Labs-blue.svg?style=flat-square)](http://ipn.io) [![](https://img.shields.io/badge/project-IPFS-blue.svg?style=flat-square)](http://ipfs.io/) [![](https://img.shields.io/badge/freenode-%23ipfs-blue.svg?style=flat-square)](http://webchat.freenode.net/?channels=%23ipfs) 
+
+> A Native Application for your OS to run your own IPFS Node. Built with Electron Shell
 
 # usage
 
@@ -9,7 +13,7 @@ npm run start
 
 # packaging
 
-will be written to ./pkg
+will be written to ./dist
 
 ## make a package for your system
