+++ app/js/menu.jsx
@@ -72,8 +72,8 @@ var Menu = React.createClass({
     ]
 
     var image = (this.state.status !== 'running'
-      ? '../node_modules/ipfs-logo/ipfs-logo-128-black.png'
-      : '../node_modules/ipfs-logo/ipfs-logo-128-ice.png')
+      ? '../../node_modules/ipfs-logo/ipfs-logo-128-black.png'
+      : '../../node_modules/ipfs-logo/ipfs-logo-128-ice.png')
 
     var uninitialized = (this.state.status === 'uninitialized')
