+++ app/views/help.html
@@ -2,13 +2,13 @@
   <head>
     <title>IPFS Setup</title>
     <meta charset="utf-8"/>
-    <link rel="stylesheet" href="css/bootstrap.min.css"/>
-    <link rel="stylesheet" href="css/common.css"/>
+    <link rel="stylesheet" href="../../node_modules/bootstrap/dist/css/bootstrap.css"/>
+    <link rel="stylesheet" href="../styles/common.css"/>
   </head>
   <body>
     <div id="header" class="row">
       <div class="col-xs-2">
-        <img id="logo" src="img/logo.png"/>
+        <img id="logo" src="../img/logo.png"/>
       </div>
       <div class="col-xs-7">
       </div>
@@ -25,6 +25,6 @@
         </div>
       </div>
     </div>
-    <script src="../js/help.js"></script>
+    <script src="build/js/help.js"></script>
   </body>
 </html>
