+++ app/views/initialize.html
@@ -1,0 +1,20 @@
+<html>
+  <head>
+    <title>IPFS Setup</title>
+    <meta charset="utf-8"/>
+    <link rel="stylesheet" href="../../node_modules/bootstrap/dist/css/bootstrap.css"/>
+    <link rel="stylesheet" href="../styles/common.css"/>
+    <script src="../js/initialize.js"></script>
+  </head>
+  <body>
+    <div id="header" class="row">
+      <div class="col-xs-2">
+        <img id="logo" alt="ipfs logo" src="../img/logo.png"/>
+      </div>
+      <div class="col-xs-7">
+      </div>
+    </div>
+    <br/>
+    <div id="initform"></div>
+  </body>
+</html>
