+++ app/views/menu.html
@@ -1,0 +1,28 @@
+<html>
+  <head>
+    <meta charset="utf-8"/>
+    <title>Example App</title>
+    <link rel="stylesheet" href="../../node_modules/bootstrap/dist/css/bootstrap.css"/>
+    <link rel="stylesheet" href="../../node_modules/bootstrap-toggle/css/bootstrap2-toggle.css"/>
+    <link rel="stylesheet" href="../styles/menu.css"/>
+    <script src="../js/menu.js"></script>
+  </head>
+  <body>
+    
+    
+<!--    <div id="error_row" class="row">
+      <div class="col-xs-9 col-xs-offset-2">
+        <div class="panel panel-danger">
+          <div class="panel-heading">
+            DAVID TEST
+          </div>
+          <div class="panel-body" id="error">
+          </div>
+        </div>
+      </div>
+    </div>
+-->
+
+    <div id="menu-app"/>
+  </body>
+</html>
