--- html/menu.html
@@ -1,13 +1,0 @@
-<html>
-  <head>
-    <meta charset="utf-8"/>
-    <title>Example App</title>
-    <link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap.css"/>
-    <link rel="stylesheet" href="../node_modules/bootstrap-toggle/css/bootstrap2-toggle.css"/>
-    <link rel="stylesheet" href="../css/menu.css"/>
-    <script src="../build/menu.js"></script>
-  </head>
-  <body>
-    <div id="menu-app"/>
-  </body>
-</html>
