+++ package.json
@@ -17,10 +17,6 @@
     "react": "0.13.1",
     "silence-chromium": "2.0.0"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ipfs/electron-app"
-  },
   "devDependencies": {
     "browserify": "10.2.3",
     "browserify-shim": "3.8.8",
@@ -30,7 +26,7 @@
     "reactify": "1.1.1",
     "rimraf": "2.3.4",
     "shelljs": "0.4.0",
-    "standard": "3.3.2"
+    "standard": "^5.0.0"
   },
   "browserify": {
     "transform": [
@@ -60,17 +56,27 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "electron": "./build.sh && node_modules/.bin/electron index.js",
     "start": "./build.sh && node_modules/.bin/electron index.js 2>&1 | node_modules/.bin/silence-chromium",
-    "lint": "git diff --name-only --cached --relative | egrep .js$ | xargs --no-run-if-empty standard",
-    "package": "./build.sh && node pkg.js",
-    "package-all": "./build.sh && node pkg.js --all"
+    "lint": "./node_modules/.bin/standard .",
+    "dist": "./build.sh && node pkg.js",
+    "dist-all": "./build.sh && node pkg.js --all"
   },
   "pre-commit": [
     "lint"
   ],
+  "standard": {
+    "ignore": [
+      "build"
+    ]
+  },
   "keywords": [
     "ipfs"
   ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/ipfs/electron-app"
+  },
   "author": "Kristoffer Ström <kristoffer@rymdkoloni.se>",
   "license": "MIT",
   "bugs": {
